requirements.txt
fastapi>=0.95.2
uvicorn>=0.15.0
pydantic>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
markdown>=3.3.0
python-frontmatter>=1.0.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
tiktoken>=0.5.0
tenacity>=8.0.1
tqdm>=4.66.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
unstructured>=0.10.30
networkx>=3.0
slowapi>=0.1.9
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
psutil>=5.9.0
# For enhanced graph RAG
regex>=2023.0.0 